@@ -22,7 +22,7 @@
         }
     ],
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "illuminate/support": "^10.0",
         "illuminate/database": "^10.0",
         "illuminate/events": "^10.0"
@@ -30,8 +30,8 @@
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.3",
         "php-parallel-lint/php-parallel-lint": "^1.2",
-        "phpunit/phpunit": "^9.5.20",
-        "lychee-org/phpstan-lychee": "^1.0",
+        "phpunit/phpunit": "^10.5",
+        "lychee-org/phpstan-lychee": "^1.0.4",
         "nunomaduro/larastan": "^2.0",
         "orchestra/testbench": "^8.0"
     },
@@ -65,7 +65,7 @@
     "prefer-stable": true,
     "config": {
         "platform": {
-            "php": "8.1"
+            "php": "8.2"
         },
         "preferred-install": "dist",
         "sort-packages": true,
