@@ -4,6 +4,9 @@ namespace Kalnoy\Nestedset;
 
 use Illuminate\Database\Eloquent\Model;
 
+/**
+ * @disregard P1037
+ */
 class DescendantsRelation extends BaseRelation
 {
 	/**
@@ -31,25 +34,20 @@ class DescendantsRelation extends BaseRelation
 	}
 
 	/**
-	 * @param Model $model
-	 * @param $related
+	 * @param Model&Node $model
+	 * @param Node       $related
 	 *
-	 * @return mixed
+	 * @return bool
 	 */
-	protected function matches(Model $model, $related)
+	protected function matches(Model $model, $related): bool
 	{
 		return $related->isDescendantOf($model);
 	}
 
 	/**
-	 * @param $hash
-	 * @param $table
-	 * @param $lft
-	 * @param $rgt
-	 *
-	 * @return string
+	 * {@inheritdoc}
 	 */
-	protected function relationExistenceCondition($hash, $table, $lft, $rgt)
+	protected function relationExistenceCondition(string $hash, string $table, string $lft, string $rgt): string
 	{
 		return "{$hash}.{$lft} between {$table}.{$lft} + 1 and {$table}.{$rgt}";
 	}
